aim for 7h30, not 8h
The 8h figure is for an employee who goes to the office in person, and
takes a 30 minute to 1h lunch break in the middle of the day.

I am sometimes working strange schedules in which I take a break in the
middle of the day to go watch a presentation at the university, in which
case I might eat on the way there. This makes it easier to include the
lunch break as part of the hours that I do not work, rather than the
hours I do work.

Therefore, I think it makes more sense to include a lunch break every
day and to aim for 7h30 instead of 8h. I sometimes take 5 minutes to
warm up food and then eat while working, in those cases I do count the
time I spend working while eating as work time, not as a lunch break.